# api/docs/swagger.yaml
basePath: /v1
definitions:
...
  entities.Message:
    properties:
      attachments:
        example:
        - https://example.com/image.jpg
        - https://example.com/video.mp4
        items:
          type: string
        type: array
      contact:
        example: "+18005550100"
        type: string
      content:
        example: This is a sample text message
        type: string
      created_at:
        example: "2022-06-05T14:26:02.302718+03:00"
        type: string
      delivered_at:
        example: "2022-06-05T14:26:09.527976+03:00"
        type: string
      encrypted:
        example: false
        type: boolean
      expired_at:
        example: "2022-06-05T14:26:09.527976+03:00"
        type: string
      failed_at:
        example: "2022-06-05T14:26:09.527976+03:00"
        type: string
      failure_reason:
        example: UNKNOWN
        type: string
      id:
        example: 32343a19-da5e-4b1b-a767-3298a73703cb
        type: string
      last_attempted_at:
        example: "2022-06-05T14:26:09.527976+03:00"
        type: string
      max_send_attempts:
        example: 1
        type: integer
      order_timestamp:
        example: "2022-06-05T14:26:09.527976+03:00"
        type: string
      owner:
        example: "+18005550199"
        type: string
      received_at:
        example: "2022-06-05T14:26:09.527976+03:00"
        type: string
      request_id:
        example: 153554b5-ae44-44a0-8f4f-7bbac5657ad4
        type: string
      request_received_at:
        example: "2022-06-05T14:26:01.520828+03:00"
        type: string
      scheduled_at:
        example: "2022-06-05T14:26:09.527976+03:00"
        type: string
      scheduled_send_time:
        example: "2022-06-05T14:26:09.527976+03:00"
        type: string
      send_attempt_count:
        example: 0
        type: integer
      send_time:
        description: SendDuration is the number of nanoseconds from when the request
          was received until when the mobile phone send the message
        example: 133414
        type: integer
      sent_at:
        example: "2022-06-05T14:26:09.527976+03:00"
        type: string
      sim:
        allOf:
        - $ref: '#/definitions/entities.SIM'
        description: |-
          SIM is the SIM card to use to send the message
          * SMS1: use the SIM card in slot 1
          * SMS2: use the SIM card in slot 2
          * DEFAULT: used the default communication SIM card
        example: DEFAULT
      status:
        example: pending
        type: string
      type:
        example: mobile-terminated
        type: string
      updated_at:
        example: "2022-06-05T14:26:10.303278+03:00"
        type: string
      user_id:
        example: WB7DRDWrJZRGbYrv2CKGkqbzvqdC
        type: string
    required:
    - attachments
    - contact
    - content
    - created_at
    - encrypted
    - id
    - max_send_attempts
    - order_timestamp
    - owner
    - request_received_at
    - send_attempt_count
    - sim
    - status
    - type
    - updated_at
    - user_id
    type: object
  entities.MessageSendSchedule:
    properties:
      created_at:
        example: "2022-06-05T14:26:02.302718+03:00"
        type: string
      id:
        example: 32343a19-da5e-4b1b-a767-3298a73703cb
        type: string
      name:
        example: Business Hours
        type: string
      timezone:
        example: Europe/Tallinn
        type: string
      updated_at:
        example: "2022-06-05T14:26:10.303278+03:00"
        type: string
      user_id:
        example: WB7DRDWrJZRGbYrv2CKGkqbzvqdC
        type: string
      windows:
        items:
          $ref: '#/definitions/entities.MessageSendScheduleWindow'
        type: array
    required:
    - created_at
    - id
    - name
    - timezone
    - updated_at
    - user_id
    - windows
    type: object
  entities.MessageSendScheduleWindow:
    properties:
      day_of_week:
        example: 1
        type: integer
      end_minute:
        example: 1020
        type: integer
      start_minute:
        example: 540
        type: integer
    required:
    - day_of_week
    - end_minute
    - start_minute
    type: object
...
  entities.SIM:
    enum:
    - SIM1
    - SIM2
    type: string
    x-enum-varnames:
    - SIM1
    - SIM2
...
  requests.MessageSendScheduleStore:
    properties:
      name:
        type: string
      timezone:
        type: string
      windows:
        items:
          $ref: '#/definitions/requests.MessageSendScheduleWindow'
        type: array
    required:
    - name
    - timezone
    - windows
    type: object
  requests.MessageSendScheduleWindow:
    properties:
      day_of_week:
        type: integer
      end_minute:
        type: integer
      start_minute:
        type: integer
    required:
    - day_of_week
    - end_minute
    - start_minute
    type: object
...
  responses.BadRequest:
    properties:
      data:
        example: The request body is not a valid JSON string
        type: string
      message:
        example: The request isn't properly formed
        type: string
      status:
        example: error
        type: string
    required:
    - data
    - message
    - status
    type: object
...
  responses.InternalServerError:
    properties:
      message:
        example: We ran into an internal error while handling the request.
        type: string
      status:
        example: error
        type: string
    required:
    - message
    - status
    type: object
...
  responses.MessageSendScheduleResponse:
    properties:
      data:
        $ref: '#/definitions/entities.MessageSendSchedule'
      message:
        example: Request handled successfully
        type: string
      status:
        example: success
        type: string
    required:
    - data
    - message
    - status
    type: object
  responses.MessageSendSchedulesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/entities.MessageSendSchedule'
        type: array
      message:
        example: Request handled successfully
        type: string
      status:
        example: success
        type: string
    required:
    - data
    - message
    - status
    type: object
...
  responses.MessagesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/entities.Message'
        type: array
      message:
        example: Request handled successfully
        type: string
      status:
        example: success
        type: string
    required:
    - data
    - message
    - status
    type: object
  responses.NoContent:
    properties:
      message:
        example: action performed successfully
        type: string
      status:
        example: success
        type: string
    required:
    - message
    - status
    type: object
  responses.NotFound:
    properties:
      message:
        example: cannot find message with ID [32343a19-da5e-4b1b-a767-3298a73703ca]
        type: string
      status:
        example: error
        type: string
    required:
    - message
    - status
    type: object
...
  responses.PaymentRequired:
    properties:
      message:
        example: You have reached the maximum number of allowed resources. Please
          upgrade your plan.
        type: string
      status:
        example: error
        type: string
    required:
    - message
    - status
    type: object
...
  responses.Unauthorized:
    properties:
      data:
        example: Make sure your API key is set in the [X-API-Key] header in the request
        type: string
      message:
        example: You are not authorized to carry out this request.
        type: string
      status:
        example: error
        type: string
    required:
    - data
    - message
    - status
    type: object
  responses.UnprocessableEntity:
    properties:
      data:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      message:
        example: validation errors while handling request
        type: string
      status:
        example: error
        type: string
    required:
    - data
    - message
    - status
    type: object
...
host: api.httpsms.com
info:
  contact:
    email: support@httpsms.com
    name: support@httpsms.com
  description: Use your Android phone to send and receive SMS messages via a simple
    programmable API with end-to-end encryption.
  license:
    name: AGPL-3.0
    url: https://raw.githubusercontent.com/NdoleStudio/http-sms-manager/main/LICENSE
  title: httpSMS API Reference
  version: "1.0"
paths:
...
  /integration/3cx/messages:
    post:
      consumes:
      - application/json
      description: Sends an SMS message from the 3CX platform
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/responses.NoContent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Unauthorized'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.UnprocessableEntity'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.InternalServerError'
      summary: Sends a 3CX SMS message
      tags:
      - 3CXIntegration
...
  /messages:
    get:
      consumes:
      - application/json
      description: Get list of messages which are sent between 2 phone numbers. It
        will be sorted by timestamp in descending order.
      parameters:
      - default: "+18005550199"
        description: the owner's phone number
        in: query
        name: owner
        required: true
        type: string
      - default: "+18005550100"
        description: the contact's phone number
        in: query
        name: contact
        required: true
        type: string
      - description: number of messages to skip
        in: query
        minimum: 0
        name: skip
        type: integer
      - description: filter messages containing query
        in: query
        name: query
        type: string
      - description: number of messages to return
        in: query
        maximum: 20
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.MessagesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Unauthorized'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.UnprocessableEntity'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.InternalServerError'
      security:
      - ApiKeyAuth: []
      summary: Get messages which are sent between 2 phone numbers
      tags:
      - Messages
...
  /send-schedules:
    get:
      description: List all send schedules owned by the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.MessageSendSchedulesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Unauthorized'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.InternalServerError'
      security:
      - ApiKeyAuth: []
      summary: List send schedules
      tags:
      - SendSchedules
    post:
      consumes:
      - application/json
      description: Create a new send schedule for the authenticated user.
      parameters:
      - description: Payload of new send schedule.
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.MessageSendScheduleStore'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.MessageSendScheduleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Unauthorized'
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/responses.PaymentRequired'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.UnprocessableEntity'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.InternalServerError'
      security:
      - ApiKeyAuth: []
      summary: Create send schedule
      tags:
      - SendSchedules
  /send-schedules/{scheduleID}:
    delete:
      description: Delete a send schedule owned by the authenticated user.
      parameters:
      - description: Schedule ID
        in: path
        name: scheduleID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Unauthorized'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.InternalServerError'
      security:
      - ApiKeyAuth: []
      summary: Delete send schedule
      tags:
      - SendSchedules
    put:
      consumes:
      - application/json
      description: Update a send schedule owned by the authenticated user.
      parameters:
      - description: Schedule ID
        in: path
        name: scheduleID
        required: true
        type: string
      - description: Payload of updated send schedule.
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.MessageSendScheduleStore'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.MessageSendScheduleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.Unauthorized'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.NotFound'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.UnprocessableEntity'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.InternalServerError'
      security:
      - ApiKeyAuth: []
      summary: Update send schedule
      tags:
      - SendSchedules
...
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-api-Key
    type: apiKey
swagger: "2.0"